Remove old scheduler.hh/cc
